requirements.txt
async-timeout==3.0.1
discord.py==1.3.2

